requirements.txt
Django~=4.1
djangorestframework~=3.13.1
django-debug-toolbar~=3.5.0
django-cors-headers~=3.13.0
drf-spectacular~=0.23.1
psycopg2~=2.9.3
PyJWT~=2.4.0
python-dotenv~=0.20.0
pydantic~=1.9.1
